Watch for errors when baking.
We need to listen for an error event or else it escalates and an
uncatchable error gets thrown. This was happening in the backpack when
someone would try to send in an image that was not a PNG   a signature
error would get emitted, but the `baker.bake` method was not listening
on the `error` event.